ai-module/requirements.txt
setuptools>=68.2.0
python-dotenv>=1.0.1
numpy==1.26.4
pandas==2.1.1
pillow==10.4.0
tensorflow==2.15.1
scipy==1.14.1
